algomethod-submissions/pyproject.toml
[tool.poetry]
authors = ["kagemeka <kagemeka1@gmail.com>"]
description = ""
license = "MIT"
name = "algomethod-submissions"
packages = [
  {include = "src"}, # {include = "pseudo", from = "src"},
]
version = "0.1.0"

[tool.poetry.dependencies]
PyYAML = "^"
algomethod = "0.1.3"
filesystem-python = "0.2.0"
python = ">=3.9,<3.11"

[tool.poetry.dev-dependencies]
autopep8 = "^"
bandit = "^1.7.2"
black = "^"
blacken-docs = "^"
flake8 = "^"
isort = "^"
mypy = "^"
myst-parser = "^"
pdoc3 = "^"
pre-commit = "^"
pydocstyle = "^"
pytest = "^"
pytest-asyncio = "^"
sphinx = "^"
sphinxcontrib-mermaid = "^"
types-requests = "^"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.extras]
[tool.black]
# compatible configurations -> https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
line-length = 79
target-version = ["py38", "py39", "py310"]

[tool.flake8]
count = 'true' # have not adaped to boolean value yet.
exclude = ""
ignore = 'E203'
max_line_length = 79
show-source = 'true'
statistics = 'true'

[tool.isort]
# all config files -> https://pycqa.github.io/isort/docs/configuration/config_files.html
# full options -> https://pycqa.github.io/isort/docs/configuration/options.html
profile = "black"
# see all profiles -> https://pycqa.github.io/isort/docs/configuration/profiles.html
line_length = 79

[tool.mypy]
allow_untyped_globals = false
color_output = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = false
disallow_untyped_decorators = true
disallow_untyped_defs = true
error_summary = true
ignore_missing_imports = true
no_implicit_optional = true
no_implicit_reexport = true
no_strict_optional = false
pretty = true
python_version = "3.10"
show_absolute_path = false
show_column_numbers = true
show_error_codes = true
strict = true
strict_equality = true
strict_optional = true

[tool.pytest.ini_options]
cache_dir = "/tmp/.pytest_cache"
